improved merge sort

+++ Algorithms/Introduction-to-Algorithms/sorting/sorting_algorithm.hpp
@@ -9,13 +9,13 @@ class SortingAlgorithm
 public:
   virtual ~SortingAlgorithm() = default;
   virtual void Sort(int* array_to_sort, int number_of_elements) = 0;
-  int NumberOfComparisons() const;
-  int NumberOfAssignements() const;
+  long long NumberOfComparisons() const;
+  long long NumberOfAssignements() const;
   void ResetCounters();
 
 protected:
-  int number_of_comparisons_ = 0;
-  int number_of_assignements_ = 0;
+  long long number_of_comparisons_ = 0;
+  long long number_of_assignements_ = 0;
 };
 
 } // namespace sorting
